Avoid using page instance to passing key info back and forth
