+++ presto-hive/pom.xml
@@ -279,6 +279,11 @@
             <artifactId>jackson-annotations</artifactId>
         </dependency>
 
+        <dependency>
+            <groupId>com.mapd</groupId>
+            <artifactId>ciderJNI</artifactId>
+        </dependency>
+
         <!-- for testing -->
         <dependency>
             <groupId>com.facebook.presto</groupId>
