+++ presto-hive/src/main/java/com/facebook/presto/hive/orc/OrcSelectivePageSource.java
@@ -15,6 +15,9 @@ package com.facebook.presto.hive.orc;
 
 import com.facebook.presto.common.InvalidFunctionArgumentException;
 import com.facebook.presto.common.Page;
+import com.facebook.presto.common.block.Block;
+import com.facebook.presto.common.block.IntArrayBlock;
+import com.facebook.presto.common.block.LongArrayBlock;
 import com.facebook.presto.hive.FileFormatDataSourceStats;
 import com.facebook.presto.orc.OrcAggregatedMemoryContext;
 import com.facebook.presto.orc.OrcCorruptionException;
@@ -22,9 +25,19 @@ import com.facebook.presto.orc.OrcDataSource;
 import com.facebook.presto.orc.OrcSelectiveRecordReader;
 import com.facebook.presto.spi.ConnectorPageSource;
 import com.facebook.presto.spi.PrestoException;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.mapd.CiderJNI;
+import sun.misc.Unsafe;
 
 import java.io.IOException;
 import java.io.UncheckedIOException;
+import java.lang.reflect.Field;
+import java.util.ArrayList;
+import java.util.Iterator;
+import java.util.List;
+import java.util.Optional;
 
 import static com.facebook.presto.hive.HiveErrorCode.HIVE_BAD_DATA;
 import static com.facebook.presto.hive.HiveErrorCode.HIVE_CURSOR_ERROR;
@@ -87,7 +100,7 @@ public class OrcSelectivePageSource
             if (page == null) {
                 close();
             }
-            return page;
+            return getOutputFromCider(page, recordReader.getSubQuery(), recordReader.getTableColumns());
         }
         catch (InvalidFunctionArgumentException e) {
             closeWithSuppression(e);
@@ -105,6 +118,151 @@ public class OrcSelectivePageSource
             closeWithSuppression(e);
             throw new PrestoException(HIVE_CURSOR_ERROR, format("Failed to read ORC file: %s", orcDataSource.getId()), e);
         }
+    }
+
+    private Page getOutputFromCider(Page page, String subQuery, String tableColumn)
+    {
+        int positionCount = page.getPositionCount();
+        int count = page.getChannelCount();
+
+        // FIXME columns??
+        int resultCount = recordReader.getColumnNum();
+        long[] dataBuffers = new long[count];
+        long[] dataNulls = new long[count];
+        long[] resultBuffers = new long[resultCount];
+        long[] resultNulls = new long[resultCount];
+
+        ObjectMapper mapper = new ObjectMapper();
+        JsonNode node = null;
+        try {
+            node = mapper.readTree(tableColumn);
+        }
+        catch (JsonProcessingException e) {
+            e.printStackTrace();
+        }
+        JsonNode columnArrays = node.get("Columns");
+        Iterator<JsonNode> nodes = columnArrays.elements();
+        List<String> list = new ArrayList<>();
+        while (nodes.hasNext()) {
+            JsonNode next = nodes.next();
+            Iterator<String> keys = next.fieldNames();
+            while (keys.hasNext()) {
+                String key = keys.next();
+                String value = next.get(key).textValue();
+                list.add(value);
+            }
+        }
+
+        Page output = null;
+        long ptr = CiderJNI.init();
+        try {
+            Unsafe unsafe = getUnsafe();
+            for (int i = 0; i < count; i++) {
+                if (list.get(i).equals("INT")) {
+                    int[] inputBuffer = ((IntArrayBlock) page.getBlock(i)).values;
+                    long nullPtr = unsafe.allocateMemory(positionCount);
+                    long bufferPtr = unsafe.allocateMemory(positionCount * 4);
+                    for (int idx = 0; idx < inputBuffer.length; idx++) {
+                        unsafe.putInt(bufferPtr + idx * 4, inputBuffer[idx]);
+                    }
+                    dataBuffers[i] = bufferPtr;
+                    dataNulls[i] = nullPtr;
+                }
+                else if (list.get(i).equals("LONG")) {
+                    long[] inputBuffer = ((LongArrayBlock) page.getBlock(i)).values;
+                    long nullPtr = unsafe.allocateMemory(positionCount);
+                    long bufferPtr = unsafe.allocateMemory(positionCount * 8);
+                    for (int idx = 0; idx < inputBuffer.length; idx++) {
+                        unsafe.putLong(bufferPtr + idx * 8, inputBuffer[idx]);
+                    }
+                    dataBuffers[i] = bufferPtr;
+                    dataNulls[i] = nullPtr;
+                }
+            }
+
+            for (int i = 0; i < resultCount; i++) {
+                // FIXME: hardcode now
+                String type = "INT";
+                int step = 0;
+                switch (type) {
+                    case "INT":
+                        step = 4;
+                        break;
+                    case "LONG":
+                        step = 8;
+                        break;
+                }
+                long nullPtr = unsafe.allocateMemory(positionCount);
+                long bufferPtr = unsafe.allocateMemory(positionCount * step);
+                resultBuffers[i] = bufferPtr;
+                resultNulls[i] = nullPtr;
+            }
+
+            int resultSize = CiderJNI.processBlocks(
+                    ptr,
+                    subQuery,
+                    tableColumn,
+                    dataBuffers,
+                    dataNulls,
+                    resultBuffers,
+                    resultNulls,
+                    positionCount);
+
+            Block[] outputBlocks = new Block[resultCount];
+            for (int i = 0; i < resultCount; i++) {
+                String type = list.get(i);
+                int step = 0;
+                switch (type) {
+                    case "INT":
+                        step = 4;
+                        break;
+                    case "LONG":
+                        step = 8;
+                        break;
+                }
+                long nullPtr = resultNulls[i];
+                long resultPtr = resultBuffers[i];
+                int index = 0;
+                if (type.equals("INT")) {
+                    int[] outputBuffer = new int[resultSize];
+                    for (int idx = 0; idx < positionCount; idx++) {
+                        // FIXME: need to confirm how nullPtr is represented
+                        if (unsafe.getByte(nullPtr + idx) == 0x01) {
+                            outputBuffer[index++] = unsafe.getInt(resultPtr + idx * step);
+                        }
+                    }
+                    outputBlocks[i] = new IntArrayBlock(resultSize, Optional.empty(), outputBuffer);
+                }
+                else if (type.equals("LONG")) {
+                    long[] outputBuffer = new long[resultSize];
+                    for (int idx = 0; idx < positionCount; idx++) {
+                        if (unsafe.getByte(nullPtr + idx) == 0x01) {
+                            outputBuffer[index++] = unsafe.getLong(resultPtr + idx * step);
+                        }
+                    }
+                    outputBlocks[i] = new LongArrayBlock(resultSize, Optional.empty(), outputBuffer);
+                }
+            }
+
+            output = new Page(resultSize, outputBlocks);
+        }
+        catch (NoSuchFieldException e) {
+            e.printStackTrace();
+        }
+        catch (IllegalAccessException e) {
+            e.printStackTrace();
+        }
+
+        CiderJNI.close(ptr);
+        return output;
+    }
+
+    private static Unsafe getUnsafe() throws NoSuchFieldException, SecurityException, IllegalArgumentException, IllegalAccessException
+    {
+        Field f = Unsafe.class.getDeclaredField("theUnsafe");
+        f.setAccessible(true);
+        Unsafe unsafe = (Unsafe) f.get(null);
+        return unsafe;
     }
 
     @Override
